refactor: Phase 11 — use-case layer re-architecture
- Move pipeline files to core/llm/pipeline/ (pipelineContext, steps, postPromptOrchestrator)
- Split 3 scope-creep files into 7 focused modules:
  chatLateEnrichment → vectorLoreInjection + memoryMessageInjection
  chatPromptShared → promptConfigReaders + promptWorkerLifecycle + promptPayloadBuilder
  memoryBookContext → memoryEmbeddingIndex + memoryKeyMatching + memoryContextInjection
- Fix naming: remove chat- prefix from generic names (4 files renamed)
- Eliminate hollow entrypoints: calculateContext, generateSummary, generateMemoryDraft
  now own dependency assembly locally instead of proxying to generationService
- generationService.js reduced from 267 → 172 lines (only exports generateChatResponse)
- Document remaining UI leak sites in ARCHITECTURE.md for Phase 12
