+++ src/core/llm/transport/completionsClient.js
